+++ src/lib.rs
@@ -199,7 +199,7 @@
 //! Example usage:
 //!
 //! ```rust
-//! # use egui_dock::{DockState, TabContextMenuTranslations, Translations, LeafTranslations};
+//! # use egui_dock::{AccessibilityTranslations, DockState, TabContextMenuTranslations, Translations, LeafTranslations};
 //! # type Tab = ();
 //! let translations_pl = Translations {
 //!     tab_context_menu: TabContextMenuTranslations {
@@ -219,6 +219,14 @@
 //!         minimize_button_menu_hint: "Kliknij prawym przyciskiem myszy, aby zminimalizować to okno.".to_string(),
 //!         minimize_button_modifier_hint: "Naciśnij klawisze modyfikujące (domyślnie Shift), aby zminimalizować to okno.".to_string(),
 //!         minimize_button_modifier_menu_hint: "Naciśnij klawisze modyfikujące (domyślnie Shift) lub kliknij prawym przyciskiem myszy, aby zminimalizować to okno.".to_string(),
+//!     },
+//!     accessibility: AccessibilityTranslations {
+//!         close_button: "Zamknij zakładkę".to_string(),
+//!         add_button: "Dodaj zakładkę".to_string(),
+//!         close_all_button: "Zamknij wszystkie zakładki".to_string(),
+//!         collapse_button: "Zwiń węzeł".to_string(),
+//!         expand_button: "Rozwiń węzeł".to_string(),
+//!         expand_window_button: "Rozwiń okno".to_string(),
 //!     }
 //! };
 //! let dock_state = DockState::<Tab>::new(vec![]).with_translations(translations_pl);
@@ -239,6 +247,12 @@
 //! dock_state.translations.leaf.minimize_button_menu_hint = "右クリックでウィンドウを最小化する".to_string();
 //! dock_state.translations.leaf.minimize_button_modifier_hint = "修飾キー（デフォルトではShift）を押すと、このウィンドウが最小化されます".to_string();
 //! dock_state.translations.leaf.minimize_button_modifier_menu_hint = "修飾キー（デフォルトではShift）を押すか、右クリックしてこのウィンドウを最小化する".to_string();
+//! dock_state.translations.accessibility.close_button = "タブを閉じる".to_string();
+//! dock_state.translations.accessibility.add_button = "タブを追加".to_string();
+//! dock_state.translations.accessibility.close_all_button = "すべてのタブを閉じる".to_string();
+//! dock_state.translations.accessibility.collapse_button = "ノードを折りたたむ".to_string();
+//! dock_state.translations.accessibility.expand_button = "ノードを展開".to_string();
+//! dock_state.translations.accessibility.expand_window_button = "ウィンドウを展開".to_string();
 //! ```
 
 #![warn(missing_docs)]